# scridddhubconstructionsoftware/backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: ScridddHub API
  version: 0.1.0
  description: >
    Source of truth for the TypeScript client generated into packages/api-client.
    Covers only what's implemented so far (Level 1: Land Parcels vertical slice).
servers:
  - url: http://localhost:8080
paths:
...
  /certification-packets/{packetID}/engineer-draft:
    put:
      operationId: upsertEngineerDraft
      summary: Set the Engineer's cost-incurred draft (Screen 8.4.1)
      parameters:
        - $ref: "#/components/parameters/PacketId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                cost_by_tower:
                  type: array
                  items:
                    $ref: "#/components/schemas/TowerCost"
                committed_not_reflected_rupees:
                  type: integer
                  format: int64
                  nullable: true
                source_note:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CertificationPacket"
        "404":
          $ref: "#/components/responses/NotFound"
  /certification-packets/{packetID}/engineer-draft/sign:
    post:
      operationId: signEngineerDraft
      summary: Engineer's legal certification of the cost-incurred draft
      parameters:
        - $ref: "#/components/parameters/PacketId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CertificationPacket"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /certification-packets/{packetID}/architect-certificate/certify:
    post:
      operationId: certifyArchitect
      summary: Architect's legal certification of % completion — refused unless a site visit was already recorded
      parameters:
        - $ref: "#/components/parameters/PacketId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [completion_pct, signed_by]
              properties:
                completion_pct:
                  type: number
                signed_by:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CertificationPacket"
        "400":
          $ref: "#/components/responses/BadRequest"
  /certification-packets/{packetID}/ca-draft:
    put:
      operationId: upsertCADraft
      summary: Set the CA's escrow reconciliation draft (Screen 8.4.2)
      parameters:
        - $ref: "#/components/parameters/PacketId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                collected_from_buyers_rupees:
                  type: integer
                  format: int64
                required_escrow_pct:
                  type: number
                actually_routed_to_escrow_rupees:
                  type: integer
                  format: int64
                source_note:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CertificationPacket"
        "404":
          $ref: "#/components/responses/NotFound"
  /certification-packets/{packetID}/ca-draft/sign:
    post:
      operationId: signCADraft
      summary: CA's legal certification of the escrow reconciliation draft
      parameters:
        - $ref: "#/components/parameters/PacketId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CertificationPacket"
        "400":
          $ref: "#/components/responses/BadRequest"
...
components:
  parameters:
...
    PacketId:
      name: packetID
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
...
    SignRequest:
      type: object
      required: [signed_by]
      properties:
        signed_by:
          type: string
    TowerCost:
      type: object
      properties:
        tower_name:
          type: string
        amount_rupees:
          type: integer
          format: int64
    DraftStatus:
      type: string
      enum: [draft, signed]
    EngineerDraft:
      type: object
      properties:
        id:
          type: string
          format: uuid
        cost_by_tower:
          type: array
          items:
            $ref: "#/components/schemas/TowerCost"
        total_incurred_rupees:
          type: integer
          format: int64
        committed_not_reflected_rupees:
          type: integer
          format: int64
          nullable: true
        source_note:
          type: string
        status:
          $ref: "#/components/schemas/DraftStatus"
        signed_by:
          type: string
        signed_at:
          type: string
          format: date-time
          nullable: true
    ArchitectCertificateStatus:
      type: string
      enum: [awaiting_site_visit, certified]
    ArchitectCertificate:
      type: object
      properties:
        id:
          type: string
          format: uuid
        completion_pct:
          type: number
          nullable: true
        site_visit_scheduled_at:
          type: string
          format: date-time
          nullable: true
        site_visit_completed_at:
          type: string
          format: date-time
          nullable: true
        status:
          $ref: "#/components/schemas/ArchitectCertificateStatus"
        signed_by:
          type: string
        signed_at:
          type: string
          format: date-time
          nullable: true
    CADraft:
      type: object
      properties:
        id:
          type: string
          format: uuid
        collected_from_buyers_rupees:
          type: integer
          format: int64
        required_escrow_pct:
          type: number
        required_to_escrow_rupees:
          type: integer
          format: int64
        actually_routed_to_escrow_rupees:
          type: integer
          format: int64
        routing_compliant:
          type: boolean
        source_note:
          type: string
        status:
          $ref: "#/components/schemas/DraftStatus"
        signed_by:
          type: string
        signed_at:
          type: string
          format: date-time
          nullable: true
    CertificationPacket:
      type: object
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        period_year:
          type: integer
        period_quarter:
          type: integer
        sent_to_professionals_at:
          type: string
          format: date-time
          nullable: true
        engineer_draft:
          $ref: "#/components/schemas/EngineerDraft"
        architect_certificate:
          $ref: "#/components/schemas/ArchitectCertificate"
        ca_draft:
          $ref: "#/components/schemas/CADraft"
        withdrawal_eligible:
          type: boolean
...